Add OS switch in UAA pre-start patch
In case a different stemcell is used, for example Ubuntu, some patches are no
longer valid for specific BOSH releases. There was an issue that the patch for
the CA certificate location failed to work on Ubuntu based images.

Add operating system switch into the UAA pre-start script using the existing
patch for that file. Depending on the operating system, different system paths
will be used.