Bump minikube to v1.6.2
The currently configured version of `minikube` in `kubecf` is a bit
outdated and newer versions contain updates to better deal with setup
problems due to VPN or network issues. For example, the current version
just hangs if there is a problem setting up a virtual machine on a macOS
host with `hyperkit` when there is interference with a VPN tool.

Bump `minikube` version to the latest available one.