fix: credhub_setup: use hard-coded UAA endpoint
On a production-ish cluster, the external route (uaa.<DOMAIN>) might be
fronted by an ingress controller / etc. with a cert signed by a CA that
is nowhere within the cluster (either a public CA, or a corp-wide CA).
This makes it intractable to try to identify the UAA server via the CC
/v2/info endpoint.  As there is no suitable BOSH link for this
elsewhere, just hard-code the server instead for now.