Remove max in flight and add every in the concourse pipeline
we introduced max-in-flight hoping to limit the number of clusters we
run at the same time. It turns out it doesn't work that way. As soon as
a deploy job is done the next one will be queued although no cluster has
been deleted. This will allow for as many clusters as we have time to
deploy until the first cluster is deleteted.

We also add the "version: every" setting to make sure we don't skip
versions if jobs complete in random order. See also here:

https://github.com/concourse/concourse/issues/736

Signed-off-by: Dimitris Karakasilis <DKarakasilis@suse.com>