feat: implement helm dependencies repository rule
The dependencies repository rule is suited for fetching and caching Helm
dependencies declared in the requirements.yaml file. Those dependencies
are also known as subcharts and are fetched as .tgz files under the
charts/ directory. The rule exports a filegroup that selects all the
charts to be used with the package rule.