feat: support eirinix as external helm chart
This makes eirinix work using an external helm chart instead of as part
of eirini-bosh-release.  We want this to move towards using the eirini
(helm) release instead.

The various eirini-extensions pre-render scripts are removed as those
BOSH jobs no longer exist, and therefore patching them will fail.

We also need to configure CC and UAA to allow access as their service
names (`api` and `uaa`), so that the ssh-proxy pod can talk to it
without having access to BOSH DNS name resolution.