feat: script to generate commented values.yaml
This adds a script that can be used to generate a YAML file where
everything is commented out, except for things marked as REQUIRED.
Also, things that are marked as HIDDEN will not be in the generated
output.

The plan is to use this to generate a version of values.yaml that can be
placed in the chart bundle (next to the chart) so that the user can use
that to start off deployment, ensuring that they don't attempt keep
defaults over different versions of the chart where they did not intend
to do so.

(cherry picked from commit d7622e8d2c0b971735e9ecb6f9efa44d54cf38b8)