fix: Provide upgrade path for ccdb encryption keys
Previous releases always prefixed the user-provided key labels
with a constant prefix of "ccdb_key_label_". This has been changed
to make it possible to import data that had used an encryption
key label not using this prefix (e.g. an SCF data export).

However, this change means that a simple upgrade from a previous
kubecf release no breaks. The default config is:

ccdb:
  encryption:
    rotation:
      key_labels:
      - encryption_key_0
      current_key_label: encryption_key_0

The naïve fix would be to manually apply the prefix during the upgrade:

ccdb:
  encryption:
    rotation:
      key_labels:
      - ccdb_key_label_encryption_key_0
      current_key_label: ccdb_key_label_encryption_key_0

By itself this still doesn't help because the old key *value* is still
stored under the old secret name. To accomodate this, the new code will
not apply the prefix to the secret name if the key label already starts
with it.

The release notes for the next kubecf release must mention that the
`ccdb.encryption.rotation` labels must be manually prefixed for the upgrade
to succeed. This is the case even if the default settings from values.yaml
are being used.

Furthermore, you cannot use both a prefixed and unprefixed key label at the
same time, like `ccdb_key_label_my_key` and `my_key` because they would both
store the value in the same secret. So don't do that!

(cherry picked from commit d9c643d28b0fc8a3b8940481d0852c9a3b148d28)