fix: ensure patches are applied
If the pod / containers go away unexpectedly (e.g. when a given node
dies), we can get into a situation where the volumes holding the
rendered job configs persist and things go bad for us:

1. Node goes away, the pod restarts (keeping volumes)
2. `cf-operator util template-render` re-renders the templates
   (This clobbers our patched changes)
3. The patch script runs, seeing the sentinel file (because the volume
   managed to survive), skipping the patching.

This fixes things by storing the digest of the _patched_ file in the
sentinel, so that when this happens we see that the patch target has an
unexpected checksum and we re-patch.

Fixes #852

(cherry picked from commit 2d60dd426b16f6d6719ea6f219332ef4e88ba783)