Delete pvc disks associated with the GKE cluster in CI jobs
Needed, since storageclasses and pvcs of the clusters have `reclaimPolicy:
Delete `, and the docs say that the disks corresponding to the pvcs default to
delete, but doesn't seem to be true. See:
  https://cloud.google.com/kubernetes-engine/docs/concepts/persistent-volumes
  https://cloud.google.com/sdk/gcloud/reference/compute/instances/set-disk-auto-delete

The gcloud account needs the following permissions:
 - Required 'compute.disks.list' permission for 'projects/foo'
 - Required 'compute.regions.list' permission for 'projects/foo'

Another implementation could be to label the cluster on creation and operate on
that.