Release branch names are release-x.y and not vx.y.z
The vx.y.z format is used for tags and should be kept separate from
the names for branches to avoid confusion once the branch and tag
no longer point to the same commit.

The release-x.y naming already matches the original intend, as show
in the image used by release.md. There is no separate release branch
for patch releases; they all live on the release branch for their
minor release (also as shown on the image).