Add suse_default_stack feature flag
It switches the default_stack from cflinuxfs3 to sle15.

https://github.com/cloudfoundry-incubator/kubecf/pull/945#issuecomment-635624549
specifies the expected functionality:

=== Diego ===

* suse_buildpacks=false and suse_default_stack=false

SUSE buildpacks are not installed, but the SUSE rootfs still is. We
can discuss if the rootfs should be removed as well in the future, but
in v2.2.x it has to stay for backwards compatibility.

However, buildpacks can be added back later via cf create-buildpack,
but stacks cannot, so maybe both stacks should always be available.

* suse_buildpacks=true and suse_default_stack=false (current default)

SUSE buildpacks would be installed, but not used by default because
the default stack remains cflinuxfs3.

* suse_buildpacks=false and suse_default_stack=true

This doesn't make much sense because a stack without buildpacks
doesn't work. I would suggest to throw an error because the
probability of user error is very high in this case.

* suse_buildpacks=true and suse_default_stack=true (default in v2.3+)

SUSE buildpacks installed and used by default.

===  Eirini ===

* suse_buildpacks=false and suse_default_stack=false

Not supported because there is no cflinuxfs3 Eirini image. Throws an error.

* suse_buildpacks=true and suse_default_stack=false (current default)

Not supported because there is no cflinuxfs3 Eirini image. In v2.2.x
it silently makes the SUSE stack the default. In v2.3+ it should throw
an error.

* suse_buildpacks=false and suse_default_stack=true

This doesn't make much sense because a stack without buildpacks
doesn't work. I would suggest to throw an error because the
probability of user error is very high in this case.

* suse_buildpacks=true and suse_default_stack=true (default in v2.3+)

SUSE buildpacks installed and used by default.