Use a separate semver resource for each deploy job
because the creation of a version is not an atomic operation and it
could happen that 2 deploy jobs use the same version. When one of them
fails it will trigger a deletion of the cluster although the other job
is still running. This results if failed jobs which are hard to explain.

Also add missing inputs in job