[gh#kubecf-738] Make cfdot available in a properly shared location.
Patched cfdot job prestart to a shared location accessible in all the container of an instance group. The patches are all the same for the three affected instance groups. (As the job itself is the same).

Added comments to the diego-cell patch setup as a reminder to maintainers as to the intent of the complexity.