Load the config file in a datasource on pipeline.yaml.gomplate
We define a variable containing the context from the "config" datasource. It
would be possible to load the config as a '.' context, but the context is lost
when in "range" or "with", and would need to be addressed with $.key, which
seems less clear than $config.key as we have with the datasource.