[gh#kubecf-505][gh#kubecf-504] Drop the patches which remove `ulimit` calls from a pre-start and various other scripts.
With the exception of (`router/gorouter/jobs/patch_pre-start.sh`) none of the scripts they change are used. They are old-style start/stop control scripts ignored by the operator (which uses the BPM information instead).

The pre-start script TOH is used, however as the change happens in a different container (versus the job runtime) it is not doing anything useful, either way.