[gh#kubecf-819] Disable/clear storage class setting for eirini persi broker.
By not specifying an SC we can make use of the DefaultSC admission plugin
to get something sensible in a more complex deployment with many helm charts
and sub charts.

Updates the local chart to new release of the eirini extensions.

__Attention__

This PR reverts PR #1118 which removed the PR #1077, the original
introduction of the functionality. PR #1118 was necessary because PR
system.

This reversion here makes sure to only re-introduce the functionality
of PR #1077 again, without losing the ssh-proxy work from PR #1027.