fix: CI: attempt to force remove terraform locks
Terraform has a locking system to avoid different invocations of it from
clobbering each other and causing issues.  As we only even run one
instance at a time, we should just get rid of it before attempting to
clean up.  Otherwise we can end up with failures to clean up when the
provisioning step fails / was aborted halfway.