fix: scheduler: Fix readiness probe
The default log output format for cc-deployment-updater has been changed
to include fractions-of-a-second; this is not supported by the default
jq conversion.  Manually delete the fractional seconds and parse the
date instead to fix the functionality.

Still using tax + grep to avoid having to slurp the whole file as one
giant array.