fix: install cflinuxfs3 buildpacks correctly
The buildpacks had hard-coded version numbers in the file names in a
place that's totally disconnected from the release version mapping; this
caused them to be missed when bumping the cf-deployment.  Use more logic
to be able to pull the versions out of the standard place instead, to
ensure that we can pick them up correctly next time too.