Implement configurable multi-stack support
This commit implements the configurable multi-stack support specified in
[Proposal: Configurable multi-stack support · Issue #1086](
https://github.com/cloudfoundry-incubator/kubecf/issues/1086).

Stacks can be defined with a single YAML config file; they no longer
require any custom logic in any of the chart files. cflinuxfs and
sle15 are defined in `config/cflinuxfs3.yaml` and `config/sle15`
respectively. There is one more reference to the sle15 stack in
`config/eirini.yaml` to configure which rootfs is included in the
`bits-service` container image.

The details of the stack support are documented in `doc/stacks.md`.

The stack implementation is localized to `assets/operations/stacks.yaml`
and `templates/_stacks.tpl`.

There are several kubecf changes that were required to make the stack
support purely configuration driven, so additional stacks can be
supplied without changing the kubecf helm chart at all:

* A new config system storing internal config data under `config/`.
  It is implemented in `templates/_config.tpl` and also provides access
  to the cf-deployment manifest.

* The config system provides a condition evaluator that can be used
  by other components, to move template logic into data structures.

* A simplified `releases` management, as the `stacks` definition has
  an extended version of the `releases` schema as a property. This is
  implemented in `assets/operations/releases.yaml` and
  `templates/_releases.tpl`. It uses the new `_config.condition`
  mechanism to configure which releases to include in the deployment.

* A `deprecations` system to catch the use of unsupported options, or
  combinations of unsupported settings, and can show a helpful message
  about the replacement mechanism for the deprecated settings. It
  uses the new condition mechanism to specify the deprecation rules.

* A validation schema for `values.yaml`. This was motivated by the
  desire to add various assertions inside the new templates. Moving
  the validation to helm itself is more comprehensive, easier to
  maintain, and provides better error messages than `{{ fail }}` in
  templates.