Don't delete credhub CA from cflinuxfs3-rootfs-setup job
We replace the job property definitions in the stacks.yaml code
and don't include the credhub cert unless credhub is enabled. So
this ops file would have removed the UAA cert when credhub is
disabled.

There is a similarly dangerous modification of the diego-rep CA
certs left, but I have no idea yet how to make it more robust.