Don't overwrite lifecycle_bundles hash completely
Just add entries for all installed stacks. That way we don't have to
duplicate the docker lifecycle setting, and won't loose any other
existing settings from cf-deployment (not that it matters; they are
all for Windows stacks, but it is still cleaner).

Also create a droplet_destination for each installed stack. Everything
seems to work fine even without it, but maybe we are missing some use-case
where this setting is used/required.