fix: chart: public svcs: don't select on version
In the public services (used when we're not using ingress), don't have a
version label in the selector when finding backends.  This is necessary
to ensure that, during upgrades, we can select for whichever pod happens
to be running at the time (including the old version).  Otherwise we'll
unncessarily introduce downtime as we refuse to connect to the perfectly
fine pods (from the previous version).

The same logic also applies to downgrades (where the working pods are
from a newer version).

For https://github.com/cloudfoundry-incubator/kubecf/issues/1257 - but
this might not fix the whole thing (as the database is still not
properly HA).