Revert "fix: Remove also the git part from the version of the artifacts"
This reverts commit 994e34f780f7211e26a176163cc1085edc469958.

That commit was added because we mangled the semver for 2.5.0 and included a git
hash in it. Hence, for releasing the tar, we removed the `g<hash>` part too. We
should have not done that, and the job was failing correctly.

Now, with 2.5.1, the semver doesn't contain the git hash, so we shouldn't try to
remove it. This is the correct behaviour.