fix: chart: Set memory limit for CCDB key rotation
We were failing in CI when attempting to rotate the CCDB keys; this
turns out to be because were were getting OOMKilled.  It appears that
the job needs more than 128Mi to complete (as in, 192Mi appears to work;
intermediate values were not tested).  Set the limits for the process
explicitly to hopefully fix CI.