fix: set diego-cell ephemeral PVC disk size
When we're using PVCs for ephemeral disks for diego-cell, we will need
to set the disk size.  The default size of 10Gi is too small to hold the
default (grootfs) disk size.  Calculate it based on the user-specified
ephemeral disk size, so that it can adjust according to user needs.

It needs to be slightly larger than the user specified size, as it is
also used to hold other internal data; as an initial estimate, make it
5Gi larger than requested.  We may be able to adjust this down later, as
for initial deployment we appear to only need <300Mi extra.