fix: Reserve space from user input rather than add
When we are using a PVC to manage ephemeral disks on diego-cell,
subtract the space in reserve from the user-provided value rather than
adding to it.  During upgrades we are unlikely to be able to resize
PVCs, so we should instead maintain the same size and just let diego use
less of it instead.

Also change the reserved space to be 2Gi instead; that should still be
plenty, but not reduce the user-visible deployable app estimate as much.