Move jobs between instance groups in $.kubecf.manifest
This way we don't need to extract the job definitions during build
time and import them again; everything can be modified in-tree, and
in the end just the modified manifest is written to the configmap.

It felt silly to create an update.serial setting for newly created
instance groups, just to delete it again in the sequencing.yaml
file via ops, so the whole handling of update.serial has been
simplified as well.