Remove blobstore settings from eirini-helm ops file
These settings are not needed, and they will break when kubecf
is configured to use an external blobstore because then the
singleton-blobstore service doesn't even exist.

I suspect that more settings in this file are redundant, but
want to have a minimal change to just fix Eirini with external
blobstore deployment.