feat: CI: Add workflow to label queued PRs
This adds a new GitHub Actions workflow, running on a timer.  Every 15
minutes, it will clean up CI-generated clusters that do not have a
matching CI run active (the assumption being that the previous run
failed and the cluster has been orphaned).  Additionally, the workflow
will also determine the number of clusters we have room for, and trigger
the correct number of queued pull requests (by marking them with a
different label).