feat: CI: Move linting to separate workflow
Since we do not need any external resources to do linting, run it on
every PR rather than waiting for approval; this leads to faster cycle
times for everybody (and ensures we don't approve things that will fail
linting).

Also switch to actually running tests on the pr-test-trigger label,
which is now automatically set by the pull-request-queue workflow.