fix: CI: Set up GCLoud config directory
If CATS takes a long time, the GCloud access token can time out; in that
case, we need to let the gcloud binary know how to find the
configuration to get a new one.  This is normally covered by catapult;
so we need to pull in enough of its configs to do so.