Use a bogus rfc5737 example IP address if there's no tcp-router svc
Concourse CI fails on the gcloud DNS transaction, as there's no tcp-router-ip
for the service.

Concourse CI doesn't use Catapult nor external-dns for the creation of GKE
clusters.

To not overcomplicate things for now, give a valid but bogus value to
`$tcp_router_ip` on the pipeline, when the service is not present. The correct
way would be to either use external-dns properly (moving the pipeline code to
Catapult) or read `features.routing_api` which is not there until later, but we
are moving to GH Actions anyways.