fix: CI: PR: run on PR branch, not master
The `pull_request_target` event actually makes us check out the target
branch (i.e. `master`), which isn't useful for gating PRs.  We'll have
to rely on people only attaching the proper label after review instead.

Also, as a precaution, drop the credentials in the checkouts; we don't
need those for the run, as we never need to push things back up in this
workflow.