feat: CI: Add CCDB rotation tests
This adds the test where we rotate the cloud controller secrets and
expect things to continue working.  This includes changing the existing
test script (aimed at concourse-based CI) to accept an existing chart,
as we no longer build the chart as part of the testing job.  This would
also make it easier to use it to test other existing builds (e.g. for
the release process).