Add an `upgrade-tests` job stub
This is a copy-paste of the `test` job, with the same job configuration and
strategy matrix. This means that it will grab 2 clusters, 1 for Diego and 1 for
Eirini, and test upgrades on them.

It's a stub, as it is missing steps for obtaining/creating the needed bundles,
and the upgrade tests.

There is repetition with the `tests` job. For the GHA boilerplate code this may
be ok. For shellscripts inside, I don't think it is ok, they need to be moved to
kubecf/scripts, catapult, or wherever it corresponds.

GHA doesn't support yaml anchors yet nor a way to not repeat steps. If we want
to remove all repetition, we should run the yaml through templating.