Avoid generating ops that reference instance groups that don't exist
Right now these are just property settings for "api" and "cc-worker"
that are set in chart/config/api.yaml and chart/config/jobs.yaml,
but this commit pre-emptively handles all groups that are removed
in chart/templates/multiple_cluster_mode.yaml.

Ideally this would all be handled by conditions defined in
chart/config/jobs.yaml, but the required logic becomes too complex,
so special-casing this issue seems warranted.

This commit also adds an op for brain-tests, which seems to be missing
because of an oversight.