feat: remove cc_deployment_updater.
The cc_deployment_updater job has an incorrect readiness probe;
when it is in HA, it can get jammed and never become ready (see below).
Rather than properly fixing it (by moving it into a separate instance
group and marking it as active/passive), just drop the job completely as
it does not appear to be useful at all.

Details on the readiness probe problems:

The probe was unaware that it was supposed to be active/passive, and
just checks the logs to ensure all instances are sleeping once every
five seconds.  This only happened to work on passive nodes because the
test incorrectly succeeded when an instance has never been ready
(because jq happily accepted empty input and returned success).  This
means that if a pod was ever active, then turned passive again (e.g.
because it was being restarted and another instance became active), it
would have a stale line that matched the probe and therefore start
getting marked as failed.