Move wait-for-ip code to a standalone script in pull-request-ci/
This reverts commit 519f6ae9fc21fa7fbac9c2dee162116d76d63689,
adds .github/workflows/pull-request-ci/wait-for-ip.sh
and consumes it on tests and upgrade-tests job

The code only works for services that are LB/Ingress-terminated, not those with
hardcoded IPs such as Minikube, so it can't be on kubecf-wait.sh.

Another solution would be to add information of what kind of service termination
we have to the cluster, so one can piece that together with the domain (e.g: by
using a configmap containing the service type).